Update
Added a validation for employee creation that Time cannot = 0.
Added in workload calculators.
Created 2 seperate views depending on if a manager or employee is logged in. Current issues: Time has to be inputed as minutes, need to be able to input as hours in the form and convert to minutes in the database. No calender. No recurring task creation. Workload calculator had issues if there was no time inputed for an employee, corrected with validation for time on employee creation, added time values to all seeds